Extract geocoding request to submodule
