+++ src/icewatch/geocode/__init__.py
@@ -1,0 +1,4 @@
+from icewatch.geocode.nomination import geocode_address
+
+
+__all__ = ["geocode_address"]
